requirements.txt
flask
kafka-python
seaborn